% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_covid.R
\name{download_covid}
\alias{download_covid}
\title{Download COVerAGE-DB data}
\usage{
download_covid(
  data = c("inputDB", "output_5", "output_10", "qualityMetrics"),
  temp = TRUE,
  tibble = FALSE,
  progress = TRUE,
  conflicts = "overwrite",
  recurse = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{The name of the dataset to download. Can be one of the
the following: "inputDB", "output_5", "output_10", "qualityMetrics".}

\item{temp}{Logical. Should the dataset be downloaded to a temporary directory?}

\item{tibble}{Logical. should the function return a tibble?}

\item{progress}{Passed to \code{\link[osfr:osf_download]{osfr::osf_download()}}. Logical, if TRUE
progress bars are displayed for each file transfer. Mainly useful for
transferring large files. For tracking lots of small files, setting
‘verbose = TRUE’ is more informative.}

\item{conflicts}{Passed to \code{\link[osfr:osf_download]{osfr::osf_download()}}. This determines what
happens when a file with the same name
exists at the specified destination. Can be one of the
following:
\itemize{
\item ‘"error"’ (the default): throw an error and abort the
file transfer operation.
\item ‘"skip"’: skip the conflicting file(s) and continue
transferring the remaining files.
\item ‘"overwrite"’: replace the existing file with the
transferred copy.
}}

\item{recurse}{Passed to \code{\link[osfr:osf_download]{osfr::osf_download()}}. Applies only to
OSF directories. If TRUE, a directory is fully recursed and all nested
files and subdirectories are downloaded. Alternatively, a positive number
will determine the number of levels to recurse.}

\item{verbose}{Passed to \code{\link[osfr:osf_download]{osfr::osf_download()}} Logical, indicating whether
to print informative messages about interactions with the OSF API (default FALSE).}

\item{...}{Additional named arguments passed to \code{\link[readr:read_delim]{readr::read_csv()}}}
}
\value{
By default a data frame or a tibble. The return value is controlled
by the "tibble" parameter.
}
\description{
Downloads one of the COVerAGE-DB datasets hosted on \href{https://osf.io/mpwjq/}{OSF}.
It reads the downloaded data and converts it into a data frame.
}
\details{
This function uses the OSF API to download the publicly available COVerAGE-DB
datasets. It then returns the downloaded dataset as a data frame or tibble.
There are 4 possible datasets available for download: "inputDB", "output_5",
"output_10", "qualityMetrics".
See \url{https://timriffe.github.io/covid_age/GettingStarted.html} for more
information about these datasets and the reading routine
used in this function.
}
\examples{
\dontrun{
# Basic usage
df <- download_covid("inputDB")
# Download data to the working directory.
df <- download_covid("output_5", temp = FALSE)
# If you want a tibble instead of a data frame:
df <- download_covid("output_10", tibble = TRUE)
}

}
\seealso{
\code{\link[osfr:osf_retrieve]{osfr::osf_retrieve()}} for the OSF entity retrieval function;
\code{\link[osfr:osf_download]{osfr::osf_download()}} for the downloading function; \code{\link[readr:read_delim]{readr::read_csv()}} for
the reading function.
}
\author{
Erez Shomron
}
