% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_covid.R
\name{read_covid}
\alias{read_covid}
\title{Read downloaded COVerAGE-DB datasets}
\usage{
read_covid(
  filepath,
  data = c("inputDB", "Output_5", "Output_10", "qualityMetrics"),
  return = c("data.frame", "data.table", "tibble"),
  ...
)
}
\arguments{
\item{filepath}{Character. The local file path of the downloaded dataset.}

\item{data}{The name of the dataset to download. Can be one of the
the following: "inputDB", "Output_5", "Output_10", "qualityMetrics".}

\item{return}{What should be the return type? Can be on of the following:
"data.frame", "data.table", "tibble".}

\item{...}{Additional named arguments passed to \code{\link[data.table:fread]{data.table::fread()}}}
}
\value{
By default a data frame with the uncompressed dataset.
Can be set to return either a data table or
a tibble. The return type is controlled by the 'return' parameter.
}
\description{
Reads one of the COVerAGE-DB datasets that was locally saved. The advantage
over the usual reading functions, is that the columns are ensured to have
their correct data types according to COVerAGE-DB documentation.
}
\details{
This function can be used to read a previously downloaded datasets. However,
to get the latest version, it's better to use \code{\link[=download_covid]{download_covid()}}.
Reading the complete 'inputDB' is quite memory intensive. It may fail
if \emph{R} cannot allocate enough memory for the dataset.
}
